started pre-poduction and optimization
