--- TODO.md
@@ -1,63 +1,0 @@
-# Circuits and logics to implement
-
-Reference Webisite : [tutorialspoint.com](https://www.tutorialspoint.com/digital_circuits/index.htm)
-### First release : 
-
-- [Bases Conversion](Conversion.py) 
-    - Binary to Decimal
-    - Binary to Hexadecimal
-    - Binary to Octal
-    - Decimal to Hexadecimal
-    - Decimal to Octal
-    - Decimal to Binary
-    - Octal to Hexadecimal
-    - Octal to Decimal
-    - Octal to Binary
-    - Hexadecimal to Octal
-    - Hexadecimal to Decimal
-    - Hexadecimal to Binary
-
-- [Logic Gates](Gates.py)
-    - AND gate
-    - OR gate
-    - NOT gate
-    - NAND gate
-    - NOR gate
-    - XOR gate
-    - XNOR gate
-
-- [Flip flops](FlipFlop.py) 
-    - Nand Latch
-    - Nor Latch
-    - SR flip flop
-    - D flip flop
-    - JK flip flop
-    - JK master slave flip flop
-    - T flip flop
-
-- [Combinational Logic Circuits](multiplexing.py)
-    - Multiplexer
-    - Demultiplexer
-    - Encoder
-    - Decoder
-
-- [Display terminals](seven-segment-display.py)
-    - Seven segment display
-
-- [Arithmetic Circuits](adder-subtractor.py)
-    - Half adder
-    - Full adder
-    - Half subtractor
-    - Full subtractor
-    - Comparator
-
-- [Counter](Counters.py)
-  - Ring counter
-  - Jonhson counter
-  - Ripple counter 
-
-- [Shift Registers](Shift_registers.py)
-  - SISO
-  - SIPO
-  - PIPO
-  - PISO
