+++ arithematic_circuit.py
@@ -1,8 +1,18 @@
-from main_functions import *
-from Gates import *
+'''
+arithematic_circuit\n
+type:function | class\n
+name-format: [name]\n
+half adder\n
+full adder\n
+half subtractor\n
+full subtractor\n
+comparator
+'''
 
+from __main_functions import *
+from gates import *
 
-
+'''Half adder'''
 def half_adder(val1,val2):
     if(valid_value(val1,val2)):
         val1,val2 = normalize_logic_values(val1,val2)
@@ -12,13 +22,14 @@ def half_adder(val1,val2):
     else:
         raise ValueError("Invalid Values!")
 
+'''Full adder'''
 def full_adder(val1,val2,carry_in = 0):
     sum_ha, carry_ha = half_adder(val1,val2)[0],half_adder(val1,val2)[1]
     sum_fa = logic_xor(sum_ha,carry_in)
     carry_fa = logic_or(logic_and(sum_ha,carry_in),carry_ha)
     return(sum_fa,carry_fa)
 
-
+'''Half subtractor'''
 def half_subtractor(val1,val2):
     if(valid_value(val1,val2)):
         val1,val2 = normalize_logic_values(val1,val2)
@@ -28,19 +39,15 @@ def half_subtractor(val1,val2):
     else:
         raise ValueError("Invalid Values!")
 
+'''Full subtractor'''
 def full_subtractor(val1,val2,borrow_in = 0):
     diff_hs, borrow_hs = half_subtractor(val1,val2)[0], half_subtractor(val1,val2)[1]
     diff_fs = logic_xor(diff_hs, borrow_in)
     borrow_fs = logic_or(logic_and(logic_not(diff_hs),borrow_in),borrow_hs)
     return(diff_fs,borrow_fs)
 
-
-
+'''Comparator'''
 class comparator():
-    def __check_valid_comparator_data(self,level):
-        if (not isinstance(level, int) or level < 0):
-            raise ValueError("Invalid value!") # invalid level
-                
                     
     def __init__(self, level,inputno = None):
         self.level = level
@@ -67,13 +74,7 @@ class comparator():
                         if len(each) > self.level:
                             if(all(each[0:len(each) - self.level]) != 0):
                                 raise ValueError(f"Input value of input {inputno.index(each) + 1} is large")    
-                    
-                    
-                
-                    
-                        
-            
-           
+
             else:
                 raise ValueError("Input list must have only two values")
         
@@ -97,20 +98,7 @@ class comparator():
         if(int("".join(str(bit) for bit in self.inputno[0]),2) < int("".join(str(bit) for bit in self.inputno[1]),2)):
             return(1)
         else: return 0
-        
-if __name__ == "__main__":
-    comparator1 = comparator(2)
-    comparator1.comparator_input([[1,1],[1,1]])
-    print(comparator1.A_gt_B())
-    print(comparator1.A_ls_B())
-    print(comparator1.A_eq_B())
 
-    comparator2 = comparator(4)
-    comparator2.comparator_input([[1,1,1,0],[1,0,0,0]])
-    print(comparator2.A_gt_B())
-    print(comparator2.A_ls_B())
-    print(comparator2.A_eq_B())
-
